Revert "refactor: move CCoinJoinQueue from coinjoin.h to common.h to drop dependency of msg_result on the heavy header"
This reverts commit e38ffdcea3b04c33f81fa6ddfe862edcbf13356e.
