+++ src/coinjoin/common.cpp
@@ -4,9 +4,7 @@
 
 #include <coinjoin/common.h>
 
-#include <bls/bls.h>
 #include <core_io.h>
-#include <logging.h>
 #include <univalue.h>
 
 namespace CoinJoin
@@ -26,32 +24,3 @@ std::string DenominationToString(int nDenom)
 }
 
 } // namespace CoinJoin
-  //
-uint256 CCoinJoinQueue::GetSignatureHash() const
-{
-    return SerializeHash(*this, SER_GETHASH, PROTOCOL_VERSION);
-}
-uint256 CCoinJoinQueue::GetHash() const { return SerializeHash(*this, SER_NETWORK, PROTOCOL_VERSION); }
-
-bool CCoinJoinQueue::CheckSignature(const CBLSPublicKey& blsPubKey) const
-{
-    if (!CBLSSignature(Span{vchSig}, false).VerifyInsecure(blsPubKey, GetSignatureHash(), false)) {
-        LogPrint(BCLog::COINJOIN, "CCoinJoinQueue::CheckSignature -- VerifyInsecure() failed\n");
-        return false;
-    }
-
-    return true;
-}
-
-bool CCoinJoinQueue::IsTimeOutOfBounds(int64_t current_time) const
-{
-    return current_time - nTime > COINJOIN_QUEUE_TIMEOUT ||
-           nTime - current_time > COINJOIN_QUEUE_TIMEOUT;
-}
-
-[[nodiscard]] std::string CCoinJoinQueue::ToString() const
-{
-    return strprintf("nDenom=%d, nTime=%lld, fReady=%s, fTried=%s, masternode=%s",
-        nDenom, nTime, fReady ? "true" : "false", fTried ? "true" : "false", masternodeOutpoint.ToStringShort());
-}
-
