+++ src/msg_result.h
@@ -5,7 +5,8 @@
 #ifndef BITCOIN_MSG_RESULT_H
 #define BITCOIN_MSG_RESULT_H
 
-#include <coinjoin/common.h>
+#include <coinjoin/coinjoin.h>
+
 #include <protocol.h>
 #include <uint256.h>
 
@@ -49,7 +50,7 @@ struct MessageProcessingResult
     std::vector<CInv> m_inventory;
 
     //! @m_dsq will relay DSQs to connected peers
-    std::optional<CCoinJoinQueue> m_dsq;
+    std::vector<CCoinJoinQueue> m_dsq;
 
     //! @m_transactions will relay transactions to peers which is ready to accept it (some peers does not accept transactions)
     std::vector<uint256> m_transactions;
