+++ test/lint/lint-circular-dependencies.py
@@ -28,11 +28,11 @@ EXPECTED_CIRCULAR_DEPENDENCIES = (
     "chainlock/chainlock -> chainlock/signing -> llmq/signing_shares -> net_processing -> llmq/context -> chainlock/chainlock",
     "chainlock/chainlock -> chainlock/signing -> llmq/signing_shares -> net_processing -> masternode/active/context -> chainlock/chainlock",
     "chainlock/chainlock -> instantsend/instantsend -> instantsend/signing -> chainlock/chainlock",
-    "chainlock/chainlock -> chainlock/signing -> llmq/signing_shares -> net_processing -> coinjoin/walletman -> coinjoin/client -> coinjoin/coinjoin -> chainlock/chainlock",
+    "chainlock/chainlock -> llmq/quorums -> msg_result -> coinjoin/coinjoin -> chainlock/chainlock",
     "chainlock/chainlock -> validation -> chainlock/chainlock",
     "chainlock/chainlock -> validation -> evo/chainhelper -> chainlock/chainlock",
     "chainlock/signing -> llmq/signing_shares -> net_processing -> masternode/active/context -> chainlock/signing",
-    "coinjoin/common -> core_io -> evo/cbtx -> llmq/blockprocessor -> msg_result -> coinjoin/common",
+    "coinjoin/coinjoin -> instantsend/instantsend -> spork -> msg_result -> coinjoin/coinjoin",
     "coinjoin/client -> coinjoin/coinjoin -> instantsend/instantsend -> instantsend/signing -> llmq/signing_shares -> net_processing -> coinjoin/walletman -> coinjoin/client",
     "coinjoin/server -> net_processing -> coinjoin/server",
     "coinjoin/server -> net_processing -> masternode/active/context -> coinjoin/server",
